Remove useless build dependency on postgresql-devel.
While at it, depend on postgresql version starting with 8.3. We
claim to support those old releases per README anyways, and testing
shows that this is still true.

However, there's a bug in _gen_recovery_conf(), which uses quoting
not yet supported in 8.3. This will be fixed by a following commit.