Support non-Linux platforms.
This is a first shot on supporting non-Linux platforms. However,
we still rely on Bash and GNU tools to operate properly, so the
target platform is required to provide all GNU tools necessary.
Currently, Solaris and OSX are the only tested non-Linux platforms
so far and we special case operation for the following GNU tools
to be available with a 'g' prefix there. The required GNU tools
are in summary:

- gtar
- gtail
- gfind
- ggrep
- gsed

Also, Solaris and OSX don't provide the flock command line tool. To
provide locking against concurrent runs of pg_backup_ctl, use an atomic
mkdir to emulate the filesystem locking (see run_with_lock() for details).