Various fixes and improvements.
Teach check_archivedir() to test for write access.

do_rsync_basebackup() has a check to check for the relocated
tablespace directory to exist. However, we need to test wether
-T was really specified, before error out. Add additional checks
to test wether -T was specified or not and teach do_rsync_basebackup()
to also test if we have write access to the relocated tablespace
directory.

do_cleanup() mistakenly assumes the basebackup filename was
specified with the full path to the archive directory. This
assumption is plain wrong, since we usually expect to get the
basebackup filename from the user (hence the basename on do_cleanup()).
Fix that by always fully qualify the specified basebackup filename.

do_rsync_basebackup() excludes the postmaster.pid file from
the basebackup now.