Heavily rework details of commit 0f916d2c2c0ee4379f8059cd962442eec42dbec0.
The new pin and unpin actions require a lot of more leg work
to be done in all cleanup actions with a specific basebackup,
backup history file or when called without any arguments. The basic
guts of the cleanup action is relocated into a separate routine
_perform_basebackup_history_cleanup(), which is responsible to
clean all basebackups which are not pinned. Since do_cleanup()
always tries to start its deletion actions based on the first
WA segment *NOT* to delete, we also need to iteratively loop
through all candidates to delete. This will check wether they are
pinned and will move the deletion offset in the WAL archive into
the past if true.