Fix handling of cleanup command when called without argument.
Bash on Solaris 11 (which is bash 4.1.11 as of this commit) didn't
like the empty case...esac branch in do_cleanup(). I dislike the handling
of that in the past anyways, so i'm going to change that to assume
the implicit argument of 'latest' in case no argument is specified
to cleanup. With this commit, you can type "cleanup latest" or
simply "cleanup" as a synomym for the latest basebackup to keep
when doing the housekeeping.

No documentation yet.