Fix cleanup in case of partially gzipped archive.
The previous code in do_cleanup for streambackups asserted that either
all files in the archive are compressed or none.  With a mix - which
happens, e.g., when changing the archive_command on a production
cluster or using compression cronjobs on the archive itself -
pg_backup_ctl would remove the backup label file but not the
accompanying base backup.